+++ src/coordinator/users.js
@@ -14,12 +14,31 @@ import { toCamelCase, toSnakeCase } from '../utils/objects';
  */
 
 /**
+ * @typedef {Object} ContxtUserApplication
+ * @property {string} applicationId
+ * @property {string} createdAt ISO 8601 Extended Format date/time string
+ * @property {string} id
+ * @property {string} userId
+ * @property {string} updatedAt ISO 8601 Extended Format date/time string
+ */
+
+/**
  * @typedef {Object} ContxtUserRole
  * @property {string} createdAt ISO 8601 Extended Format date/time string
  * @property {string} id
  * @property {boolean} mappedFromExternalGroup
  * @property {string} userId
  * @property {string} roleId
+ * @property {string} updatedAt ISO 8601 Extended Format date/time string
+ */
+
+/**
+ * @typedef {Object} ContxtUserStack
+ * @property {string} accessType Access Type of the user for this stack with options "reader", "collaborator", "owner"
+ * @property {string} createdAt ISO 8601 Extended Format date/time string
+ * @property {string} id
+ * @property {string} userId
+ * @property {string} stackId
  * @property {string} updatedAt ISO 8601 Extended Format date/time string
  */
 
@@ -94,10 +113,49 @@ class Users {
   }
 
   /**
+   * Adds a application to a user
+   *
+   * API Endpoint: '/users/:userId/applications/:applicationId'
+   * Method: GET
+   *
+   * @param {string} userId The ID of the user
+   * @param {string} applicationId The ID of the application
+   *
+   * @returns {Promise}
+   * @fulfill {ContxtUserApplication} The newly created user application
+   * @reject {Error}
+   *
+   * @example
+   * contxtSdk.coordinator.users
+   *   .addApplication('36b8421a-cc4a-4204-b839-1397374fb16b', '007ca9ee-ece7-4931-9d11-9b4fd97d4d58')
+   *   .then((userApplication) => console.log(userApplication))
+   *   .catch((err) => console.log(err));
+   */
+  addApplication(userId, applicationId) {
+    if (!userId) {
+      return Promise.reject(
+        new Error('A user ID is required for adding a application to a user')
+      );
+    }
+
+    if (!applicationId) {
+      return Promise.reject(
+        new Error(
+          'An application ID is required for adding a application to a user'
+        )
+      );
+    }
+
+    return this._request
+      .post(`${this._baseUrl}/users/${userId}/applications/${applicationId}`)
+      .then((response) => toCamelCase(response));
+  }
+
+  /**
    * Adds a role to a user
    *
    * API Endpoint: '/users/:userId/roles/:roleId'
-   * Method: GET
+   * Method: POST
    *
    * @param {string} userId The ID of the user
    * @param {string} roleId The ID of the role
@@ -127,6 +185,54 @@ class Users {
 
     return this._request
       .post(`${this._baseUrl}/users/${userId}/roles/${roleId}`)
+      .then((response) => toCamelCase(response));
+  }
+
+  /**
+   * Adds a stack to a user
+   *
+   * API Endpoint: '/users/:userId/stacks/:stackId'
+   * Method: POST
+   *
+   * @param {string} userId The ID of the user
+   * @param {string} stackId The ID of the stack
+   * @param {'reader' | 'collaborator' | 'owner'} accessType The level of access for the user
+   *
+   * @returns {Promise}
+   * @fulfill {ContxtUserStack} The newly created user stack
+   * @reject {Error}
+   *
+   * @example
+   * contxtSdk.coordinator.users
+   *   .addStack('36b8421a-cc4a-4204-b839-1397374fb16b', '007ca9ee-ece7-4931-9d11-9b4fd97d4d58', 'collaborator')
+   *   .then((userStack) => console.log(userStack))
+   *   .catch((err) => console.log(err));
+   */
+  addStack(userId, stackId, accessType) {
+    if (!userId) {
+      return Promise.reject(
+        new Error('A user ID is required for adding a stack to a user')
+      );
+    }
+
+    if (!stackId) {
+      return Promise.reject(
+        new Error('A stack ID is required for adding a stack to a user')
+      );
+    }
+
+    if (['reader', 'collaborator', 'owner'].indexOf(accessType) === -1) {
+      return Promise.reject(
+        new Error(
+          'An access type of "reader", "collaborator", or "owner" is required for adding a stack to a user'
+        )
+      );
+    }
+
+    return this._request
+      .post(`${this._baseUrl}/users/${userId}/stacks/${stackId}`, {
+        access_type: accessType
+      })
       .then((response) => toCamelCase(response));
   }
 
@@ -294,6 +400,46 @@ class Users {
   }
 
   /**
+   * Removes a application from a user
+   *
+   * API Endpoint: '/users/:userId/applications/:applicationId'
+   * Method: DELETE
+   *
+   * @param {string} userId The ID of the user
+   * @param {string} applicationId The ID of the application
+   *
+   * @returns {Promise}
+   * @fulfill {undefined}
+   * @reject {Error}
+   *
+   * @example
+   * contxtSdk.coordinator.users
+   *   .removeApplication('36b8421a-cc4a-4204-b839-1397374fb16b', '007ca9ee-ece7-4931-9d11-9b4fd97d4d58')
+   *   .catch((err) => console.log(err));
+   */
+  removeApplication(userId, applicationId) {
+    if (!userId) {
+      return Promise.reject(
+        new Error(
+          'A user ID is required for removing a application from a user'
+        )
+      );
+    }
+
+    if (!applicationId) {
+      return Promise.reject(
+        new Error(
+          'An application ID is required for removing a application from a user'
+        )
+      );
+    }
+
+    return this._request.delete(
+      `${this._baseUrl}/users/${userId}/applications/${applicationId}`
+    );
+  }
+
+  /**
    * Removes a role from a user
    *
    * API Endpoint: '/users/:userId/roles/:roleId'
@@ -326,6 +472,42 @@ class Users {
 
     return this._request.delete(
       `${this._baseUrl}/users/${userId}/roles/${roleId}`
+    );
+  }
+
+  /**
+   * Removes a stack from a user
+   *
+   * API Endpoint: '/users/:userId/stacks/:stackId'
+   * Method: DELETE
+   *
+   * @param {string} userId The ID of the user
+   * @param {string} stackId The ID of the stack
+   *
+   * @returns {Promise}
+   * @fulfill {undefined}
+   * @reject {Error}
+   *
+   * @example
+   * contxtSdk.coordinator.users
+   *   .removeStack('36b8421a-cc4a-4204-b839-1397374fb16b', '007ca9ee-ece7-4931-9d11-9b4fd97d4d58')
+   *   .catch((err) => console.log(err));
+   */
+  removeStack(userId, stackId) {
+    if (!userId) {
+      return Promise.reject(
+        new Error('A user ID is required for removing a stack from a user')
+      );
+    }
+
+    if (!stackId) {
+      return Promise.reject(
+        new Error('A stack ID is required for removing a stack from a user')
+      );
+    }
+
+    return this._request.delete(
+      `${this._baseUrl}/users/${userId}/stacks/${stackId}`
     );
   }
 }